+++ src/templates/payment/paymentscript.js
