+++ src/templates/vehicle/vehiclescript.js
