+++ src/utils/dateUtils.js
@@ -1,0 +1,23 @@
+function parseDate(dateString) {
+    const date = new Date(dateString);
+    return date.toISOString().slice(0, 16);
+}
+
+function formatDate(dateString) {
+    const date = new Date(dateString);
+
+    const day = String(date.getUTCDate()).padStart(2, '0');
+    const month = String(date.getUTCMonth() + 1).padStart(2, '0');
+    const year = date.getUTCFullYear();
+
+    const hours = String(date.getUTCHours()).padStart(2, '0');
+    const minutes = String(date.getUTCMinutes()).padStart(2, '0');
+    const seconds = String(date.getUTCSeconds()).padStart(2, '0');
+
+    return `${day}.${month}.${year} ${hours}:${minutes}:${seconds}`;
+}
+
+module.exports = {
+    parseDate,
+    formatDate
+};
